Move asset loading in actor from BeginPlay to separate function
This allows us to have full control over when the asset is loaded,
instead of strictly during BeginPlay.

The new function `LoadAsset` will be called by a manager that will spawn
and control the `glTFRuntimeAssetActor`. For more information, refer to
the corresponding manager in The Forge.

Note: might affect Blueprint usage, but our workflow is currently only
on C++, and this change is to accomodate that.