fix: set specular to 0.5 by default if unspecified
Existing versions of Blender (currently up to 3.2) do not export
specular values in glTF/glb models. When Blender re-imports these
models, the specular values are 0.5 by default.

However, this value is 0 by default in Unreal Engine. To ensure that the
assets in Forge are as similar to Blender/glb as possible, set the
default specular in glTFRuntime to 0.5 as well.