feat: modify ClearCoat_gltfRuntimeBase for variation
Add changes identical to gltfRuntimeBase to ClearCoat_glTFRuntimeBase.
These changes are required for Forge to modify the material in runtime.

Materials that have a non-zero clearcoat value will be imported with
ClearCoat_glTFRuntimeBase instead of glTFRuntimeBase. We previously only
modified glTFRuntimeBase but since ClearCoat_glTFRuntimeBase does not
inherit from it and is its own parent material, the same changes should
be reflected in it.

Certain assets (e.g. cars) would have a non-zero clear coat as it adds a
thin translucent layer on top similar to adding a literal clear coat in
real life.