fix: enable compute fog per pixel for translucent materials
This fix addresses the issue where glass materials (using
`M_glTFRuntimeTranslucent_Inst`) would be fully visible through UDW
clouds. The fix is to enable "Compute Fog Per Pixel" on
`M_glTFRuntimeBase`, which is the parent material of the MI mentioned
above, since this setting is only available on Materials and not MIs.

Because "Compute Fog Per Pixel" is only available on materials with
"Blend Mode" set to "Translucent", make this change by temporarily
setting the "Blend Mode", updating the translucency properties, then
setting the "Blend Mode" back to "Opaque". This is so that other
materials can continue to have a default value of "Opaque".

Source and description of the fix can be found on Linear issue
[STA-559](https://linear.app/bifrostai/issue/STA-559/cockpit-glass-seen-through-clouds).