feat: baseColor MF includes `Foundation`and `Tint`
In order to do material randomization we need to augment base material
with exposed parameters. Therefore add the `FoundationColor` and
respective `foundationColorAlpha` and `TintColor` and `TintColorAlpha`.

This was done directly on the MaterialFunction which both
`M_glTFRuntimeBase` and `M_ClearCoat_glTFRuntimeBase` make use of.

This follows up on the work previously done in a02debf3 and 293490fd