unreal: add MF_GetWorldPositionOffset
This material function is responsible for faking off nadir skews on
actors. It functions by masking the Z height of the asset and adding a
parameterized offset `Skew`. This `Skew` parameter works on both X and Y
axis and is normalized but not clamped to original actor radius. Visual
diff screenshot will be added to associated MR description.

The use of actor radius can lead to uneven skew values between actors
this is just an assumption to simplify the user interaction (by
normalizing it to a known value). This still needs further validation in
user review. This is a non issue when using off nadir angles.

In regards to Off nadir angles, note that the implementation of off
nadir skew offset will likely change in the future to better fit the
concept of Off nadir angle. This is just the  first pass at it to allow
users to experiment and validate it.

The changes to the multiple materials is to connect newly added material
function to the `WorldPositionOffset` pin. This is applied to all
/GltfRuntime base materials and custom ThermalBase materials.