unreal: fix log spam from streaming hack
Change the streaming MIPs to use BULKDATA_PayloadAtEndOfFile instead
of BULKDATA_PayloadInSeperateFile as markers. According to the
comments, BULKDATA_PayloadInSeperateFile was used to avoid the MIPs
from being grouped under non streaming MIPs in TextureDerivedData.cpp.
This however caused a lot of warnings with the MIP being wrongly
classified as being on disk when its not (Loading non-streamed mips
from an external bulk file).

This new flag is a variant of the streaming hack but with less log
spam side effects every time an asset is loaded.